Give RegType a 'notional' TupleDescriptor
So that TupleTableSlot may be used uniformly as the API for
Java <-> PostgreSQL data type conversions, let every type except
unmodified RECORD or VOID have a 'notional' TupleDescriptor.

For a cataloged or interned row type, or a domain over a cataloged row
type, it is that type's tupleDescriptor() (or that of the transitive
base type, in the case of a domain). Such a descriptor will be of type
TupleDescriptor.Interned. Otherwise, it is a TupleDescriptor.Ephemeral
whose one, unnamed, attribute has this type.

The idea is that every language handler will see a TupleTableSlot into
which a routine's results should be stored, even if just one column is
there. It will be up to the common dispatcher code to grok the specific
PostgreSQL rules, "a scalar gets returned, but OUT parameters make
a composite result, unless it's just one OUT parameter and that's treated
just like a scalar, but a polymorphic type later resolved to a one-column
composite isn't" and so on, and return what is in the TupleTableSlot to
PostgreSQL in the proper way.

This notional descriptor is not exposed in RegType API, but only on
RegTypeImpl for internal use. It will exposed through a method on
a RegProcedure's PLJavaBased memo to get the routine's outputsTemplate.
Template, because what is computed here depends only on catalog
information, and may include polymorphic types needing resolution at
actual call sites.