Add API methods on RegProcedure and memo
Using the memo on a RegProcedure<PLJavaBased>, get a TupleDescriptor
describing the incoming parameters, the notional one describing the
expected results, and (because these are 'templates' that may include
polymorphic types that need later resolution), a BitSet for each,
indicating at which positions type resolution is needed. An empty
BitSet indicates the template will be exactly the descriptor seen
later at call sites.

Internally, these methods live on RegProcedureImpl, relying on its
SwitchPointCache slots for caching. The memo implementation simply
delegates to those. Alternatively, the implementations could be
moved there.

Naturally, these methods will not be usable until a dispatcher is
implemented that can know a PLJavaBased RegProcedure when it sees
one, and attach this memo to it.

New interfaces on RegProcedure are Lookup (roughly corresponding to the
PostgreSQL per-call-site struct, FmgrInfo, usually passed around as
flinfo), and Call (like the PostgreSQL per-call struct usually passed
around as fcinfo and I'm not going to type that long struct name here).

Where the memo on a RegProcedure can give you 'template' tuple
descriptors based only on the catalog and possibly with some unresolved
types, on Lookup you find methods returning the same descriptors with
types all resolved according to the types at the call site. From Call,
of course, you can get the argument and result TupleTableSlots and fetch
the argument values (and store the results someday, when that direction
gets implemented for TupleTableSlot).

These interfaces, too, await a dispatcher that will supply instances of
them to your code.