Interfaces a PL/Java-based language implements
Here is the interface PLJavaBasedLanguage, with its two subinterfaces
InlineBlocks and Routines (any PLJavaBasedLanguage must additionally
implement either one or both of those). PostgreSQL's CREATE LANGUAGE
allows the inline handler to be optional, while the handler for
routines is mandatory. PL/Java won't mind, though, if a language
only implements InlineBlocks; CREATE LANGUAGE will still have to
mention pljavaDispatchRoutine, but only inline blocks will really
be allowed.

A staged-programming idiom is the approach for Routines. The prepare
method is invoked passing only the target RegProcedure, and should
return a Template that depends only on information available from
that RegProcedure and its PLJavaBased memo. The Template will be
cached with the RegProcedure itself.

At a new call site, the Template's specialize method will be applied
to the call site's Lookup instance, where it can refer to
call-site-specific information like the fully resolved argument and
result types, returning a Routine to be cached for as long as PostgreSQL
has not freed that call site. (Naturally, when there is nothing
polymorphic and no need to specialize, the Routine can be constructed
all at prepare() time, and the Template can just unconditionally return
it.

For every call through a given call site, its cached Routine's call
method is applied to a Call instance that will supply the arguments
and (one day) accept the results.