Add DualState.SingleDeleteGlobalRefP
The caching of a Lookup object at a PostgreSQL call site (in
flinfo->fn_extra) will call for a flavor of DualState that can
delete a JNI global reference when the call site's memory context
is reset or deleted.

This seems to be the first DualState flavor whose nativeStateReleased
method actually does anything, which requires a little refinement of
some only-when-assertions-enabled checking that had never had to
deal with that before.