Now put the implementation under it
STILL UNIMPLEMENTED AT THIS STAGE: anything to do with context
classloader management or access control contexts. A language's
trusted bit or name don't matter yet; the code will run with
whatever is granted to the intersection of its and PL/Java's
codebases without any Principal-based grants. Also,
Function_currentLoader, Function_currentTypeMap, and
Function_isCurrentReadOnly won't yet work when the call didn't
go through the legacy dispatcher, which may well cause errors
or crashes in attempting to use JDBC or SPI.

An internal-only field is added to ProceduralLanguageImpl's
RoutineSet as a place to memoize the implementing class instance,
and one to the PLJavaMemo as a place to memoize the Template
generated for a RegProcedure.

The Call methods context() and resultinfo() will return objects
if PostgreSQL supplied corresponding nodes, but those classes are,
so far, only stubs with no useful methods.

Call.result() is unimplemented for now and throws an exception.
A writable TupleTableSlot is still future work.

Call.isNull(boolean) works,  to determine whether the current
implementation will return null or void. For now, though, that's
unconditionally done by the dispatcher after the custom handler
returns, based on the type expected by PostgreSQL: null is
returned for any non-byValue type, to avoid an immediate crash,
and void is returned otherwise (some callers expecting void are
unexpectedly cranky if the void they get back is the null kind).
This corner of the implementation is still stopgap until writable
TupleTableSlot happens.