Fix added example for library-name diversity
Serves me right for thinking to save a bit of time by putting
the new handler function declarations into an example for now,
and deferring the work of adding them to InstallHelper and
declaring a new schema version.

But of course a hardcoded library name in the example won't
pass CI when Mac and Windows spell library names differently.
So the work is still done in an example for now, but in a
plpgsql inline block not really any tidier than doing it in Java.