No doInPG around releaseFromJava
Mechanically rewriting synchronized(THREADLOCK) as doInPG
in c3f833b was a little too mechanical where releaseFromJava
is involved. None of those DualState resource-releasing methods
bother with the stack-base-fudging business, and that isn't a bug:
they are not supposed to need to. DualState is already designed
to enqueue those actions for later when the thread is known
to be right. In allow mode, the doInPG simply fools it into
performing the action immediately.

Addresses issue #471.