Back to nashorn instead of graaljs for jdk22
The version of graaljs that works on jdk15-21 won't work
on 22, and the 24.0.0 version of its js-scriptengine uses
class file version 61 and therefore won't work below 17.

As nashorn has been made available as a Maven resource
also, and works painlessly, use that for jdk22 and up.
(For later: use in place of graaljs for 15-21 also, if
tested with no issues.)