Let docs reflect upstream rename of PostgresNode
The Perl module supplied with PostgreSQL and formerly known as
PostgresNode is renamed PostgreSQL::Test::Cluster in PG 15.

This patch keeps the original naming of the Java Node.class
and simply fixes the documentation to reflect the changed name
of the upstream Perl module (and explain the historical source
of the Java name).