Restore doc buildability after site plugin bump
The version of maven-site-plugin bumped to in 5a21027
insists that the content of a mailing list <otherArchive>
be a valid URI. No more "gmane.comp.db.postgresql.pljava
on news.gmane.io".

Replace that string with a news: URI constructed according
to RFC 5538 and meaning the same thing. I have no ready access
to software that can understand and dereference such a URI,
but it ought to work, and the docs can be built again, anyway.