Prune doc references to unsupported PG versions
17b7d7c (PR #450) dropped conditional code for PostgreSQL
before 9.5, the oldest version to be supported by PL/Java 1.6.x.
References to older PostgreSQL versions in the docs, however,
were not pruned then. Remove those now.

The docs also mentioned some paths and other details that changed
with the new build system in 1.6; bring those up to date. Also,
link to the page on pljava.policy from a couple more places that
only touched on policy generally. These changes are not strictly
about pre-9.5 PostgreSQL versions, but are details that were
similarly overlooked in the PL/Java 1.5 to 1.6 transition.