Include command/arguments in Maven debug output
For the Windows cases, log the arguments before applying the
forWindowsCRuntime() transformation (but indicating, in the message,
that it will be applied).

Eliminate the debug logging of the script text in ScriptingMojo
and ReportScriptingMojo; Maven itself, when producing debug output,
will have dumped that to the log perhaps twice already (when reading
the configuration, and when configuring the mojo).