Javadoc/whitespace cleanup and slight refactoring
As dubious as combining any refactoring with a Javadoc/whitespace
cleanup may be, it was preferable to writing new Javadoc to try to
explain why it was factored as it was (with some script engine
bindings being added in PGXSUtils.getScriptEngine and others in
ScriptingMojo.execute, especially as that didn't match where the
underlying methods lived).

Adjusts javadoc options to include package-access types and members
in pljava-pgxs, as some of those are exported to the script engine
and ought to be documented.