Remove more pre-PL/Java-1.6 remnants from docs
The build system introduced in 1.6 fixed the "Windows might
get -Dpljava.libjvmdefault wrong" issue. Also, by no longer
trying to capture compiler/linker messages and classify them,
it has fixed the "Not all [ERROR]s are errors" issue.

In 1.6, both the pljava and pljava-api jars must be present
in $SHAREDIR/pljava at runtime. So a package maintainer may
choose to include the API jar additionally in a -devel package
that places it where users' development tools will expect, but
may not omit it from the base package that installs it where
the runtime needs it.