Use pgjdbc rather than pgjdbc-ng in CI
JEP 472 issues warnings (one day to be errors) for uses of
System.loadLibrary or when native methods are first bound according to
the default binding process.

PL/Java's use of JNI escapes without warnings because the JVM is
launched from C and uses RegisterNatives to bind the native methods
explicitly, so it does not involve loadLibrary or any use of the
default native method matching.

On the other hand, the pgjdbc-ng library furnishes netty, which has
a use of loadLibrary, and produces warnings from the client side
in the CI script. As Node.java now supports both pgjdbc and pgjdbc-ng,
switch to pgjdbc in the CI scripts to avoid a nuisance warning.