More, necessary, GitHub Actions workflow polishing
None of this is related at all to JEP 472, but some is needed to catch
up with changes on the GitHub runners as the CI would otherwise fail.

The macos-12 runner is gone, and it was the last macos one that supplied
PostgreSQL preinstalled. So rename the workflow from ci-runnerpg to
ci-lazypg ... it still uses the preinstalled PG on runners that have
one, but now resorts to brew to install it on macos.

The macos-13 runner is still Intel-based (and needs PG installed).
The macos-14 runner is ARM64-based (and needs PG installed).

The runners have some Java versions preinstalled, so check for those and
only fetch a JDK if the matrix.java version isn't one of the
preinstalled ones.

Preinstalled versions can be found by their JAVA_HOME_$VERSION_$ARCH
environment variables, except that when $ARCH is ARM64 the environment
variable has arm in lowercase. So much for consistency.

There are newer versions of the setup-java and checkout actions.
The old ones elicited warnings about a deprecated Node.js version.

The old setup-java, when running on macos ARM64 (or arm64), would
happily install an x86_64 JDK. Which would seem to work fine, apparently
running in a compatibility mode, apparently inherited by the C compiler
and linker spawned from Maven. The linker would then report baffling
errors about not finding symbols in PostgreSQL for x86_64! This is the
exact sort of mystery seen and not yet solved in issues #411 and #428.
So, running Maven on a JDK that's for the wrong architecture is now a
known way to cause such a problem, and worth checking in such a case.

Added some ::group:: / ::endgroup:: to make the log output more easily
perused.