Restrict trusted CREATE FUNCTION when unenforced
In the no-enforcement mode selected with java.security.manager=disallow,
enforce (in the validator) that function creation even in a PL flagged
'trusted' is reserved to superusers. This check is made unconditionally,
right after CheckFunctionValidatorAccess but before the short-circuit
return made if that returns false (a reserved possible future case),
and before consulting check_function_bodies.

Naturally, any functions already defined in 'trusted' PLs, prior to a
PL/Java and/or Java update newly configured for no enforcement, remain
defined. Selectively enabling those to be executable in no-enforcement
mode will be the subject of the next patch.