Add pljava.allow_unenforced configuration setting
The value is a comma-separated list of PL names (such as java, javau,
possibly others created with sqlj.alias_java_language). Only functions
defined in the PLs named here will be allowed to execute when the
java.security.manager=disallow property setting is in effect.

This will also be checked in the validator, thus preventing creation
of a new function if the PL is not named here. This is checked in the
validator only when check_function_bodies is on, so it is better viewed
as a reminder than as meaningful security. The check at function
invocation is the important one.