Hook JVM abort and exit
The JVM has long has hooks for both, just as sparsely documented as
the vfprintf hook, which PL/Java has already long used. This is
a good time to start hooking abort and exit also, in part because of
the seemingly ever-growing set of JVM startup issues that get reported
with an abort instead of a civilized error return from JNI_CreateJavaVM,
and in part because if there's to be no policy enforcement around
System.exit(), one likes at least to know what has happened if it gets
called. (Regrettably, the exit hook is not given the option to just
say "no".)