Defensively copy system properties
There are so many things resting on assumptions that various properties
aren't just freely modifiable by any code anywhere.

Add an unmodifiable FrozenProperties final subclass of Properties, and
populate it with a defensive copy just after InstallHelper has set the
PL/Java-specific properties. Make this frozen copy available to user
code through the Session API.

The above is not enough by itself, because PL/Java internally uses some
properties in code that executes before that point in PL/Java startup.
So those places have to be fixed to do their own defensive caching.

ELogFormatter is easy because writing a newline can be done with
PrintWriter.println(), and that uses a copy of line.separator already
cached early in the JVM startup.

The determination of the byte ordering for SQLInputFromChunk and
SQLOutputToChunk is more work. Just move all that logic out of
InstallHelper and into a new class that just serves as a cache for that,
and arrange for the new class to be initialized at the right point in
PL/Java startup.