Assume javadoc-21 form of URLs into Javadoc
While PL/Java's docs are meant to build (mvn site site:stage) without
failure on any Java version supported for building PL/Java, there have
been changes over the Javadoc versions to things like the output directory
structure and the spelling of anchor names for use as URL fragment IDs.
Therefore, it's necessary to pick a version of Javadoc to be used when
generating the docs if such things as links from the Markdown documents
into the generated javadocs are to work right.

Links from the generated javadocs to Oracle's published JDK javadocs
are made to the Java 12 pages as a practical compromise: PL/Java 1.6.x
is built for compatibility back to Java 9, but Oracle's published pages
for Java 12 are the earliest to have the organization expected by a modern
Javadoc.

The PL/Java examples (non-modular code) have Javadoc links back to the
PL/Java API (modular code), which Javadoc before 15 doesn't know how
to do. Later Javadoc versions have some further aesthetic improvements.

For now, assume that Javadoc 21 (the recent LTS version) will be used
for generating the site javadocs, and adjust the URLs that point
into the javadocs accordingly.