Likewise update releasenotes-pre1_6
A slightly Orwellian feeling is probably no important reason not to also
fix links in releasenotes-pre1_6 so they aren't broken when the docs are
built with a present-day javadoc.

There will still be unavoidable breakage given that anchor names for
functions are generated by recent Javadoc versions to include parentheses,
which are sanitized out of links by Velocity's Markdown processor.