Reformat some code
1032e68 wrapped some JavaScript code in function execute() { ... }
which was exdented, to avoid reindenting everything inside, probably
with the intention to have a later reformat-only commit making the
indentation consistent again.

This happened both in pljava-packaging/pom.xml and in pljava-so/pom.xml.
In the first, there was no code outside of what became the function body,
so the result was just to have the content of the CDATA section start
at the left margin. It looks fine. So reformat pljava-so/pom.xml the same
way, moving the lines outside the function back to the margin, rather than
further indenting the function and its body.

This reformatting is not totally mechanical: some line breaks have been
changed to stay within 80 columns, and one java.util.List.of shortened
to of, having apparently been overlooked in 9111286.