Enable CI for Windows MSVC in GitHub Actions
The Windows runner comes with an MSVC-built PostgreSQL preinstalled
(probably the Chocolatey one, which relies on the EnterpriseDB installer).

The fussiest part of configuration is finding the Visual Studio .bat file
that sets all the environment variables needed for the build to succeed,
but there is an action ilammy/msvc-dev-cmd that takes care of finding and
running that, capturing its changes to the environment, and replicating
those in the environment of the subsequent steps.

For some reason java.home on Windows has the JVM shared object under
bin\server and not lib/server as on Linux.