Also enable MinGW-w64 test with PG from pacman
MSYS2 is preinstalled on the runner but apparently not with gcc.
Have pacman install that also, and prepend the c:\msys64\mingw64\bin
directory to the PATH. Using msys2bash, it has to be prepended in
the form /c/msys64/mingw64/bin, which gets transformed correctly into
the Windows environment variable.

The gcc that is installed on the runner is in /c/mingw64/bin rather
than /c/msys64/mingw64/bin where pacman puts it. A lot of cycles and
bandwidth would be saved if that one could be used. PL/Java can be
successfully built that way, but on loading it into the backend,
"server process (PID 2116) was terminated by exception 0xC0000005".

So it seems that building to load into a pacman PostgreSQL backend
must be done using the pacman gcc.