Match implementor non-case-folded form to provider
The DDRProcessor creates an implied dependency of an 'implementor'
name on the SQLAction carrying a matching 'provides' string. An
implementor name, per ISO SQL/JRT, is an SQL identifier,
case-insensitive unless quoted. But a PL/Java 'provides' string is just
a string, with only case-sensitive exact matching. The DDRProcessor was
using a surprising and undocumented rule, matching only if the provides
string was identical to the implementor name's lower-case-folded form.

The results could be especially puzzling because the 'weak' nature of
the implementor/provides dependency meant there was no error or warning
of the failue to match. The dependency simply wasn't created, leading
possibly to an incorrectly-ordered deployment descriptor.

Rather than some more general and complicated solution, simply change
the rule so the provider string is matched to the non-folded implementor
name as it was spelled in the source annotation. Thus as long as the
programmer does the expected thing and spells them the same in the
source, the right thing happens.

Addresses #515.