Document Java stack traces and debugger control
As #508 highlighted, information on the log_min_messages or
client_min_messages settings needed for Java stack traces to be shown
has been mentioned in passing in too many places but not described in
the actual user documentation even once. Fix that.

The same new section is a reasonable place to also discuss how to attach
a Java debugger.