In Node, construct proper URI from Path
Sure enough, addressing the deprecation of URL constructors
by substituting URI.toURL() leads to failure on Windows where
Node.installSaxon used to get away with string concatenation
to make a URI from a Path.

Path.toUri() is a bit brain-damaged: it makes an absolute file: URI
even from a relative Path, and it inserts an empty authority instead
of just leaving the authority out. The latter problem is easily dealt
with; installExamples was already doing so. The former we'll put up with
for now.