Slight polishing of dispatcher
The fact that an implementation of PLJavaBasedLanguage needs a constructor
with a ProceduralLanguage parameter should be documented in the interface,
not just in example code.

I disliked the early return in LookupImpl.routineValidator when
check_function_bodies is false. It should perhaps complete silently
in that case if the language implementing class can't be found or
instantiated, but should otherwise make a best effort to call the
essentialChecks and additionalChecks methods and let them interpret
their checkBody parameter.