Make better use of memo created during validation
The process of dispatching to a validator ends up creating most of the
usual linkages among the subject RegProcedure and its language and
validator, but of course the memo attached to the RegProcedure has null
where the Template result of prepare() ought to be.

Past successful testing was a happy result of PostgreSQL shared-
invalidation events tearing down the partly-created linkages (either
when validation failed and the transaction rolled back, or even when it
committed). But that can't be counted on if creation / validation and
use follow closely within one transaction.

Two solutions: just discard the incomplete memo, as shared-invalidation
would do, and let everything be recreated on first use of the routine.
That's done now if validation was with checkBody false, as the routine
isn't known to have passed all checks.

On the other hand, if checkBody was true, it's nearly painless to just
populate the memo with a fake Template that will, when its specialize
method is later called, prepare and memoize the real Template and then
call its specialize method and return the result. Because this fake
template is planted by the validator after full validation, it can even
skip the extra call of essentialChecks that otherwise would normally
precede a call of prepare.