Use Identifier.None and TupleDescImpl.OfType more
The s_anonymous in AttributeImpl.OfType dates from 42805b1;
Identifier.None appeared later in 7faf6c9 but wasn't used here.

TupleDescImpl appeared in 42805b1 complete with an OfType subclass
intended to serve as an ephemeral descriptor representing a single
column of a given type. That seems to fit perfectly the story of
the notionalDescriptor needed when the dispatcher makes a TupleTableSlot
to receive a non-composite function result. But notionalDescriptor,
appearing in 4567117, didn't use it, instead using synthesizeDescriptor
to form an actual PostgreSQL descriptor for it. OfType is less work.
Simplifies some native code too.

It does need to override exists() though.