Big ol' rework of TargetList / Projection
PostgreSQL data structures sometimes represent a thing's
affected columns as a bitmapset or as an array of int2 (and often
using the one-based SQL convention), so it will be convenient to
have versions of project() and sqlProject() taking BitSet or short...
parameters in addition to the existing int... offerings.

That calls for an overdue refactoring effort, resisting the temptation
to just multiply the lines of code by three.

In the process, eliminate Attribute's strange foundIn method. There must
have been a day when I didn't remember that TupleDescriptor and
TargetList are Lists, and List specifies a perfectly good contains()
method for the tests foundIn was being used for.

By incorporating a BitSet into a Projection, those contains() checks
can be kept O(1). And if the Projection keeps the columns in the same
relative order as in the base TupleDescriptor, then the BitSet is all
it needs.

That does mean those cases are better based on AbstractSequentialList,
and it could be advisable to use a TargetList's iterator whenever
possible rather than making many get() calls. The whole TargetList /
Cursor design is already geared toward encouraging that anyway.