Implementation of Trigger model
The division of ModelConstants.c into sections where the macros
refer to different Java class prefixes was getting a little hairy.
Now has more structured comments so a search up for BEGIN:<macro> or
down for END:<macro> will be a quick way to determine what <macro>
refers to in a given region of the file.

Already, the lazy only-works-sometimes implementation for this model
has required special accommodation in the CatalogObjects example-cum-
regression-test.