Implementation of TriggerData model
Enough constants are getting used in LookupImpl to possibly warrant
moving them there. There's now a bit of a split: some TriggerData
(and other LookupImpl) constants are still over in ModelConstants
(which is able to read the ones that may have to be determined at
run time), while some others that may be less likely to change and
amenable to simple static checking are declared locally in LookupImpl.