Let arrays of anyarray type be manipulated
There is a case where polymorphic type resolution does not produce
resolved inputs or outputs descriptors free of unresolved types.
The type anyarray would ordinarily be seen in pre-resolution inputs
or outputs templates, and be replaced by some actual array type during
resolution. But when something is still typed anyarray after resolution,
it will be one of certain columns in statistics-related system catalogs,
or an expression derived from such a column. While definitely an array,
in another row or another call of a routine, even through the same call
site, it may be an array with a different element type.

That hinders construction of an Adapter.Array for access to the content,
since that must be built on a suitably-chosen Adapter for the element
type. Except for that difficulty, the needed type information is readily
available to an array contract, as it is passed a TupleTableSlot.Indexed
whose descriptor() reveals the element type.

So, create Adapter.Opaque, an adapter that claims applicability to all
possible types, and can't really fetch anything. It is straightforward
then to create an array adapter using Opaque as the element adapter
and a contract that simply looks at the slot descriptor and returns
the element type. Add a static method on Adapter.Array to obtain this
adapter, and document the circumstances where it would be wanted.

Discussion:
https://www.postgresql.org/message-id/2968353.1743962392%40sss.pgh.pa.us