Use a smarter SwitchPoint in SwitchPointCache
Commit 5adf2c8 raised (in its commit message) the idea of a SwitchPoint
subclass "that would set a flag, or invoke a Runnable, the first time
its guardWithTest method is called. If that hasn't happened, there is
nothing to invalidate."

This commit follows through on that idea (though instead of a Runnable,
there is an instance method onFirstUse() that can be overridden
in a subclass). The name of the subclass is still SwitchPoint to reduce
code churn to a change of an import. All existing invalidate methods
now begin with a quick return if sp.unused() is true.

The further ideas in 5adf2c8 about propagating anything-to-do status
to native flags for an even quicker return from C invalidation callbacks
remain future work, but will be straightforward to build on these changes.

Even on its own, this commit eliminates a bunch of duplicated work
resulting from the way PostgreSQL calls invalidation callbacks twice
for local changes (once right after the changes are locally visible,
and again when the backend sees its own shared-invalidation messages
after transaction end).

Noticed in passing that, even though 08e2079 commented on the rightness
of the "constructor assigns final SwitchPoint array" pattern and fixed
RegTypeImpl to match, RegClassImpl also wasn't using it and didn't
get fixed. Fixed here.

Also noticed that in many classes s_initializer was not final. Fixed.